+++ extensions/control-plane/callback/callback-event-dispatcher/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
     api(project(":spi:common:core-spi"))
+    api(project(":spi:control-plane:control-plane-spi"))
 
     testImplementation(project(":core:common:junit"))
 }
