+++ extensions/control-plane/callback/callback-event-dispatcher/src/main/java/org/eclipse/edc/connector/callback/CallbackProtocolResolverRegistryImpl.java
@@ -14,8 +14,9 @@
 
 package org.eclipse.edc.connector.callback;
 
-import org.eclipse.edc.spi.callback.CallbackProtocolResolver;
-import org.eclipse.edc.spi.callback.CallbackProtocolResolverRegistry;
+
+import org.eclipse.edc.connector.spi.callback.CallbackProtocolResolver;
+import org.eclipse.edc.connector.spi.callback.CallbackProtocolResolverRegistry;
 
 import java.util.ArrayList;
 import java.util.List;
