+++ extensions/control-plane/callback/callback-http-dispatcher/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 dependencies {
     api(project(":spi:common:core-spi"))
     api(project(":spi:common:http-spi"))
+    api(project(":spi:control-plane:control-plane-spi"))
 
 
     testImplementation(libs.mockserver.netty)
