+++ extensions/control-plane/callback/callback-http-dispatcher/src/main/java/org/eclipse/edc/connector/callback/dispatcher/http/CallbackEventRemoteMessageDispatcher.java
@@ -20,8 +20,8 @@ import okhttp3.MediaType;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
+import org.eclipse.edc.connector.spi.callback.CallbackEventRemoteMessage;
 import org.eclipse.edc.spi.EdcException;
-import org.eclipse.edc.spi.callback.CallbackEventRemoteMessage;
 
 import java.util.function.Function;
