+++ spi/control-plane/control-plane-spi/src/main/java/org/eclipse/edc/connector/spi/callback/CallbackEventRemoteMessage.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.spi.callback;
+package org.eclipse.edc.connector.spi.callback;
 
 import org.eclipse.edc.spi.event.Event;
 import org.eclipse.edc.spi.event.EventEnvelope;
