+++ spi/control-plane/control-plane-spi/src/main/java/org/eclipse/edc/connector/spi/callback/CallbackProtocolResolverRegistry.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.spi.callback;
+package org.eclipse.edc.connector.spi.callback;
 
 import org.eclipse.edc.runtime.metamodel.annotation.ExtensionPoint;
 import org.eclipse.edc.spi.message.RemoteMessageDispatcher;
